feat(js): Add getUserInput()
This commit adds getUserInput() that handles getting player's input.
This will be responsible for checking if the player clicked on
'Cancel' or 'Ok'.

If the value is `null` then this will end the game immediately.
If it is `''` then the player will be informed to enter a number
before starting to play.

If nor of the two values mentioned above are true then player's input
will then be converted into int from string data type which then will
be evaluated by validateUserInput().