feat(js): Add validateUserInput()
This commit adds validateUserInput() that is responsible for testing if
the player's input is valid. Player's input must be a number only from
one and above. If it is not a number or the value is less than or equal
to zero then the player will be informed to enter a number only and
will return the getUserInput().

Otherwise, this will return `true` confirming the value is indeed
a number from one and above.